Registering NodeHealthStore in constructor instead of StoreConfig
